requirements.txt
tavily-python
google-generativeai

